Began refactoring by using standard python package structure and removing unnecessary imports.
