+++ examples/Lshape/main_Lshape_bound.py
@@ -1,15 +1,8 @@
-import sys
-sys.path.append('../../includes')
 import cupy as cp
 import numpy as np
-from cupyx import scatter_add
 import cupyx.scipy.sparse as cusparse
-import cupyx.scipy.sparse.linalg
-import pandas as pd
-import time
-from preprocessor import write_keywords,write_birth,write_parameters
-from gamma import domain_mgr, heat_solve_mgr
-from func import elastic_stiff_matrix,constitutive_problem,transformation,disp_match
+from gamma.simulator.gamma import domain_mgr, heat_solve_mgr
+from gamma.simulator.func import elastic_stiff_matrix,constitutive_problem,transformation,disp_match
 cp.cuda.Device(3).use()
 import pyvista as pv
 import vtk
