+++ src/gamma/simulator/func.py
