+++ tests/README.md
@@ -1,0 +1,3 @@
+# Tests
+
+Place future test files here.
